refactor: move test submodules into their own files
