+++ apollo-router/src/plugins/connectors/tests.rs
@@ -36,186 +36,19 @@ use crate::services::router::Request;
 use crate::services::supergraph;
 use crate::Configuration;
 
-pub(crate) mod mock_api {
-    struct PathTemplate(String);
+mod mock_api;
+mod quickstart;
+#[allow(dead_code)]
+mod req_asserts;
 
-    impl wiremock::Match for PathTemplate {
-        fn matches(&self, request: &wiremock::Request) -> bool {
-            let path = request.url.path();
-            let path = path.split('/');
-            let template = self.0.split('/');
-
-            for pair in path.zip_longest(template) {
-                match pair {
-                    EitherOrBoth::Both(p, t) => {
-                        if t.starts_with('{') && t.ends_with('}') {
-                            continue;
-                        }
-
-                        if p != t {
-                            return false;
-                        }
-                    }
-                    _ => return false,
-                }
-            }
-            true
-        }
-    }
-
-    #[allow(dead_code)]
-    fn path_template(template: &str) -> PathTemplate {
-        PathTemplate(template.to_string())
-    }
-
-    use super::*;
-
-    pub(crate) fn users() -> Mock {
-        Mock::given(method("GET")).and(path("/users")).respond_with(
-            ResponseTemplate::new(200).set_body_json(serde_json::json!([
-              {
-                "id": 1,
-                "name": "Leanne Graham"
-              },
-              {
-                "id": 2,
-                "name": "Ervin Howell",
-              }
-            ])),
-        )
-    }
-
-    pub(crate) fn user_2_nicknames() -> Mock {
-        Mock::given(method("GET"))
-            .and(path("/users/2/nicknames"))
-            .respond_with(ResponseTemplate::new(200).set_body_json(serde_json::json!(["cat"])))
-    }
-
-    pub(crate) fn users_error() -> Mock {
-        Mock::given(method("GET")).and(path("/users")).respond_with(
-            ResponseTemplate::new(404).set_body_json(serde_json::json!([
-                {
-                    "kind": "json",
-                    "content": {},
-                    "selection": null
-                }
-            ])),
-        )
-    }
-
-    pub(crate) fn user_1() -> Mock {
-        Mock::given(method("GET"))
-            .and(path("/users/1"))
-            .respond_with(ResponseTemplate::new(200).set_body_json(serde_json::json!({
-              "id": 1,
-              "name": "Leanne Graham",
-              "username": "Bret",
-              "phone": "1-770-736-8031 x56442",
-              "email": "Sincere@april.biz",
-              "website": "hildegard.org"
-            })))
-    }
-
-    pub(crate) fn user_2() -> Mock {
-        Mock::given(method("GET"))
-            .and(path("/users/2"))
-            .respond_with(ResponseTemplate::new(200).set_body_json(serde_json::json!({
-              "id": 2,
-              "name": "Ervin Howell",
-              "username": "Antonette",
-              "phone": "1-770-736-8031 x56442",
-              "email": "Shanna@melissa.tv",
-              "website": "anastasia.net"
-            })))
-    }
-
-    pub(crate) fn create_user() -> Mock {
-        Mock::given(method("POST")).and(path("/user")).respond_with(
-            ResponseTemplate::new(200).set_body_json(serde_json::json!(
-              {
-                "id": 3,
-                "username": "New User"
-              }
-            )),
-        )
-    }
-
-    pub(crate) fn user_1_with_pet() -> Mock {
-        Mock::given(method("GET"))
-            .and(path("/users/1"))
-            .respond_with(ResponseTemplate::new(200).set_body_json(serde_json::json!({
-              "id": 1,
-              "name": "Leanne Graham",
-              "pet": {
-                  "name": "Spot"
-              }
-            })))
-    }
-
-    pub(crate) fn commits() -> Mock {
-        Mock::given(method("GET"))
-            .and(path("/repos/foo/bar/commits"))
-            .respond_with(ResponseTemplate::new(200).set_body_json(serde_json::json!(
-              [
-                {
-                  "sha": "abcdef",
-                  "commit": {
-                    "author": {
-                      "name": "Foo Bar",
-                      "email": "noone@nowhere",
-                      "date": "2024-07-09T01:22:33Z"
-                    },
-                    "message": "commit message",
-                  },
-                }]
-            )))
-    }
-
-    pub(crate) fn posts() -> Mock {
-        Mock::given(method("GET")).and(path("/posts")).respond_with(
-            ResponseTemplate::new(200).set_body_json(serde_json::json!([
-              {
-                "id": 1,
-                "title": "Post 1",
-                "userId": 1
-              },
-              {
-                "id": 2,
-                "title": "Post 2",
-                "userId": 2
-              }
-            ])),
-        )
-    }
-}
-
-pub(crate) mod mock_subgraph {
-    use super::*;
-
-    pub(crate) fn user_entity_query() -> Mock {
-        Mock::given(method("POST"))
-            .and(path("/graphql"))
-            .and(body_json(json!({
-              "query": "query($representations:[_Any!]!){_entities(representations:$representations){...on User{c}}}",
-              "variables": {"representations":[{"__typename":"User","id":1},{"__typename":"User","id":2}]}
-            })))
-            .respond_with(
-                ResponseTemplate::new(200)
-                    .insert_header(CONTENT_TYPE, APPLICATION_JSON.essence_str())
-                    .set_body_json(json!({
-                      "data": {
-                        "_entities": [{
-                          "__typename": "User",
-                          "c": "1",
-                        }, {
-                          "__typename": "User",
-                          "c": "2",
-                        }]
-                      }
-                    })),
-            )
-    }
-}
+const STEEL_THREAD_SCHEMA: &str = include_str!("./testdata/steelthread.graphql");
+const MUTATION_SCHEMA: &str = include_str!("./testdata/mutation.graphql");
+const NULLABILITY_SCHEMA: &str = include_str!("./testdata/nullability.graphql");
+const SELECTION_SCHEMA: &str = include_str!("./testdata/selection.graphql");
+const NO_SOURCES_SCHEMA: &str = include_str!("./testdata/connector-without-source.graphql");
+const QUICKSTART_SCHEMA: &str = include_str!("./testdata/quickstart.graphql");
+const INTERFACE_OBJECT_SCHEMA: &str = include_str!("./testdata/interface-object.graphql");
+const VARIABLES_SCHEMA: &str = include_str!("./testdata/variables.graphql");
 
 #[tokio::test]
 async fn value_from_config() {
@@ -444,7 +277,27 @@ async fn test_root_field_plus_entity_plus_requires() {
     mock_api::users().mount(&mock_server).await;
     mock_api::user_1().mount(&mock_server).await;
     mock_api::user_2().mount(&mock_server).await;
-    mock_subgraph::user_entity_query().mount(&mock_server).await;
+    Mock::given(method("POST"))
+            .and(path("/graphql"))
+            .and(body_json(json!({
+              "query": "query($representations:[_Any!]!){_entities(representations:$representations){...on User{c}}}",
+              "variables": {"representations":[{"__typename":"User","id":1},{"__typename":"User","id":2}]}
+            })))
+            .respond_with(
+                ResponseTemplate::new(200)
+                    .insert_header(CONTENT_TYPE, APPLICATION_JSON.essence_str())
+                    .set_body_json(json!({
+                      "data": {
+                        "_entities": [{
+                          "__typename": "User",
+                          "c": "1",
+                        }, {
+                          "__typename": "User",
+                          "c": "2",
+                        }]
+                      }
+                    })),
+            ).mount(&mock_server).await;
 
     let response = execute(
         STEEL_THREAD_SCHEMA,
@@ -1469,294 +1322,6 @@ async fn test_sources_in_context() {
     );
 }
 
-mod quickstart_tests {
-    use super::*;
-
-    macro_rules! map {
-        ($($tt:tt)*) => {
-          serde_json_bytes::json!($($tt)*).as_object().unwrap().clone()
-        };
-    }
-
-    async fn execute(query: &str, variables: JsonMap) -> (serde_json::Value, MockServer) {
-        let mock_server = MockServer::start().await;
-        Mock::given(method("GET")).and(path("/posts")).respond_with(
-            ResponseTemplate::new(200).set_body_json(serde_json::json!([
-                {
-                  "userId": 1,
-                  "id": 1,
-                  "title": "sunt aut facere repellat provident occaecati excepturi optio reprehenderit",
-                  "body": "quia et suscipit\nsuscipit recusandae consequuntur expedita et cum\nreprehenderit molestiae ut ut quas totam\nnostrum rerum est autem sunt rem eveniet architecto"
-                },
-                {
-                  "userId": 1,
-                  "id": 2,
-                  "title": "qui est esse",
-                  "body": "est rerum tempore vitae\nsequi sint nihil reprehenderit dolor beatae ea dolores neque\nfugiat blanditiis voluptate porro vel nihil molestiae ut reiciendis\nqui aperiam non debitis possimus qui neque nisi nulla"
-                }]
-            )),
-        ).mount(&mock_server).await;
-        Mock::given(method("GET")).and(path("/posts/1")).respond_with(
-            ResponseTemplate::new(200).set_body_json(serde_json::json!(
-                {
-                  "userId": 1,
-                  "id": 1,
-                  "title": "sunt aut facere repellat provident occaecati excepturi optio reprehenderit",
-                  "body": "quia et suscipit\nsuscipit recusandae consequuntur expedita et cum\nreprehenderit molestiae ut ut quas totam\nnostrum rerum est autem sunt rem eveniet architecto"
-                }
-            )),
-        ).mount(&mock_server).await;
-        Mock::given(method("GET")).and(path("/posts/2")).respond_with(
-            ResponseTemplate::new(200).set_body_json(serde_json::json!({
-                  "userId": 1,
-                  "id": 2,
-                  "title": "qui est esse",
-                  "body": "est rerum tempore vitae\nsequi sint nihil reprehenderit dolor beatae ea dolores neque\nfugiat blanditiis voluptate porro vel nihil molestiae ut reiciendis\nqui aperiam non debitis possimus qui neque nisi nulla"
-                }
-            )),
-        ).mount(&mock_server).await;
-        Mock::given(method("GET"))
-            .and(path("/users/1"))
-            .respond_with(ResponseTemplate::new(200).set_body_json(serde_json::json!({
-              "id": 1,
-              "name": "Leanne Graham",
-              "username": "Bret",
-              "email": "Sincere@april.biz",
-              "address": {
-                "street": "Kulas Light",
-                "suite": "Apt. 556",
-                "city": "Gwenborough",
-                "zipcode": "92998-3874",
-                "geo": {
-                  "lat": "-37.3159",
-                  "lng": "81.1496"
-                }
-              },
-              "phone": "1-770-736-8031 x56442",
-              "website": "hildegard.org",
-              "company": {
-                "name": "Romaguera-Crona",
-                "catchPhrase": "Multi-layered client-server neural-net",
-                "bs": "harness real-time e-markets"
-              }
-            })))
-            .mount(&mock_server)
-            .await;
-        Mock::given(method("GET")).and(path("/users/1/posts")).respond_with(
-            ResponseTemplate::new(200).set_body_json(serde_json::json!([
-                {
-                  "userId": 1,
-                  "id": 1,
-                  "title": "sunt aut facere repellat provident occaecati excepturi optio reprehenderit",
-                  "body": "quia et suscipit\nsuscipit recusandae consequuntur expedita et cum\nreprehenderit molestiae ut ut quas totam\nnostrum rerum est autem sunt rem eveniet architecto"
-                },
-                {
-                  "userId": 1,
-                  "id": 2,
-                  "title": "qui est esse",
-                  "body": "est rerum tempore vitae\nsequi sint nihil reprehenderit dolor beatae ea dolores neque\nfugiat blanditiis voluptate porro vel nihil molestiae ut reiciendis\nqui aperiam non debitis possimus qui neque nisi nulla"
-                }]
-            )),
-        ).mount(&mock_server).await;
-
-        let res = super::execute(
-            &QUICKSTART_SCHEMA.replace("https://jsonplaceholder.typicode.com", &mock_server.uri()),
-            &mock_server.uri(),
-            query,
-            variables,
-            None,
-            |_| {},
-        )
-        .await;
-
-        (res, mock_server)
-    }
-
-    #[tokio::test]
-    async fn query_1() {
-        let query = r#"
-          query Posts {
-            posts {
-              id
-              body
-              title
-            }
-          }
-        "#;
-
-        let (response, server) = execute(query, Default::default()).await;
-
-        insta::assert_json_snapshot!(response, @r###"
-        {
-          "data": {
-            "posts": [
-              {
-                "id": 1,
-                "body": "quia et suscipit\nsuscipit recusandae consequuntur expedita et cum\nreprehenderit molestiae ut ut quas totam\nnostrum rerum est autem sunt rem eveniet architecto",
-                "title": "sunt aut facere repellat provident occaecati excepturi optio reprehenderit"
-              },
-              {
-                "id": 2,
-                "body": "est rerum tempore vitae\nsequi sint nihil reprehenderit dolor beatae ea dolores neque\nfugiat blanditiis voluptate porro vel nihil molestiae ut reiciendis\nqui aperiam non debitis possimus qui neque nisi nulla",
-                "title": "qui est esse"
-              }
-            ]
-          }
-        }
-        "###);
-
-        req_asserts::matches(
-            &server.received_requests().await.unwrap(),
-            vec![Matcher::new().method("GET").path("/posts").build()],
-        );
-    }
-
-    #[tokio::test]
-    async fn query_2() {
-        let query = r#"
-          query Post($postId: ID!) {
-            post(id: $postId) {
-              id
-              title
-              body
-            }
-          }
-        "#;
-
-        let (response, server) = execute(query, map!({ "postId": "1" })).await;
-
-        insta::assert_json_snapshot!(response, @r###"
-        {
-          "data": {
-            "post": {
-              "id": 1,
-              "title": "sunt aut facere repellat provident occaecati excepturi optio reprehenderit",
-              "body": "quia et suscipit\nsuscipit recusandae consequuntur expedita et cum\nreprehenderit molestiae ut ut quas totam\nnostrum rerum est autem sunt rem eveniet architecto"
-            }
-          }
-        }
-        "###);
-
-        req_asserts::matches(
-            &server.received_requests().await.unwrap(),
-            vec![Matcher::new().method("GET").path("/posts/1").build()],
-        );
-    }
-
-    #[tokio::test]
-    async fn query_3() {
-        let query = r#"
-          query PostWithAuthor($postId: ID!) {
-            post(id: $postId) {
-              id
-              title
-              body
-              author {
-                id
-                name
-              }
-            }
-          }
-      "#;
-
-        let (response, server) = execute(query, map!({ "postId": "1" })).await;
-
-        insta::assert_json_snapshot!(response, @r###"
-        {
-          "data": {
-            "post": {
-              "id": 1,
-              "title": "sunt aut facere repellat provident occaecati excepturi optio reprehenderit",
-              "body": "quia et suscipit\nsuscipit recusandae consequuntur expedita et cum\nreprehenderit molestiae ut ut quas totam\nnostrum rerum est autem sunt rem eveniet architecto",
-              "author": {
-                "id": 1,
-                "name": "Leanne Graham"
-              }
-            }
-          }
-        }
-        "###);
-
-        req_asserts::matches(
-            &server.received_requests().await.unwrap(),
-            vec![
-                Matcher::new().method("GET").path("/posts/1").build(),
-                Matcher::new().method("GET").path("/users/1").build(),
-            ],
-        );
-    }
-
-    #[tokio::test]
-    async fn query_4() {
-        let query = r#"
-          query PostsForUser($userId: ID!) {
-            user(id: $userId) {
-              id
-              name
-              posts {
-                id
-                title
-                author {
-                  id
-                  name
-                }
-              }
-            }
-          }
-      "#;
-
-        let (response, server) = execute(query, map!({ "userId": "1" })).await;
-
-        insta::assert_json_snapshot!(response, @r###"
-        {
-          "data": {
-            "user": {
-              "id": 1,
-              "name": "Leanne Graham",
-              "posts": [
-                {
-                  "id": 1,
-                  "title": "sunt aut facere repellat provident occaecati excepturi optio reprehenderit",
-                  "author": {
-                    "id": 1,
-                    "name": "Leanne Graham"
-                  }
-                },
-                {
-                  "id": 2,
-                  "title": "qui est esse",
-                  "author": {
-                    "id": 1,
-                    "name": "Leanne Graham"
-                  }
-                }
-              ]
-            }
-          }
-        }
-        "###);
-
-        req_asserts::matches(
-            &server.received_requests().await.unwrap(),
-            vec![
-                Matcher::new().method("GET").path("/users/1").build(),
-                Matcher::new().method("GET").path("/users/1/posts").build(),
-                Matcher::new().method("GET").path("/posts/1").build(),
-                Matcher::new().method("GET").path("/posts/2").build(),
-                Matcher::new().method("GET").path("/users/1").build(),
-            ],
-        );
-    }
-}
-
-const STEEL_THREAD_SCHEMA: &str = include_str!("./testdata/steelthread.graphql");
-const MUTATION_SCHEMA: &str = include_str!("./testdata/mutation.graphql");
-const NULLABILITY_SCHEMA: &str = include_str!("./testdata/nullability.graphql");
-const SELECTION_SCHEMA: &str = include_str!("./testdata/selection.graphql");
-const NO_SOURCES_SCHEMA: &str = include_str!("./testdata/connector-without-source.graphql");
-const QUICKSTART_SCHEMA: &str = include_str!("./testdata/quickstart.graphql");
-const INTERFACE_OBJECT_SCHEMA: &str = include_str!("./testdata/interface-object.graphql");
-
 async fn execute(
     schema: &str,
     uri: &str,
@@ -1827,146 +1392,4 @@ async fn execute(
         .unwrap();
 
     serde_json::from_slice(&response).unwrap()
-}
-
-#[allow(dead_code)]
-mod req_asserts {
-    use std::collections::HashMap;
-    use std::collections::HashSet;
-
-    use itertools::Itertools;
-    use wiremock::http::HeaderName;
-    use wiremock::http::HeaderValue;
-    use wiremock::http::HeaderValues;
-
-    #[derive(Clone)]
-    pub(crate) struct Matcher {
-        method: Option<String>,
-        path: Option<String>,
-        query: Option<String>,
-        body: Option<serde_json::Value>,
-        headers: HashMap<HeaderName, HeaderValues>,
-    }
-
-    impl Matcher {
-        pub(crate) fn new() -> Self {
-            Self {
-                method: None,
-                path: None,
-                query: None,
-                body: None,
-                headers: Default::default(),
-            }
-        }
-
-        pub(crate) fn method(&mut self, method: &str) -> &mut Self {
-            self.method = Some(method.to_string());
-            self
-        }
-
-        pub(crate) fn path(&mut self, path: &str) -> &mut Self {
-            self.path = Some(path.to_string());
-            self
-        }
-
-        pub(crate) fn query(&mut self, query: &str) -> &mut Self {
-            self.query = Some(query.to_string());
-            self
-        }
-
-        pub(crate) fn body(&mut self, body: serde_json::Value) -> &mut Self {
-            self.body = Some(body);
-            self
-        }
-
-        pub(crate) fn header(&mut self, name: HeaderName, value: HeaderValue) -> &mut Self {
-            let values = self.headers.entry(name).or_insert(Vec::new().into());
-            values.append(&mut Vec::from([value]).into());
-            self
-        }
-
-        pub(crate) fn build(&mut self) -> Self {
-            self.clone()
-        }
-
-        fn matches(&self, request: &wiremock::Request, index: usize) {
-            if let Some(method) = self.method.as_ref() {
-                assert_eq!(
-                    method,
-                    &request.method.to_string(),
-                    "[Request {}]: Expected method {}, got {}",
-                    index,
-                    method,
-                    request.method
-                )
-            }
-
-            if let Some(path) = self.path.as_ref() {
-                assert_eq!(
-                    path,
-                    request.url.path(),
-                    "[Request {}]: Expected path {}, got {}",
-                    index,
-                    path,
-                    request.url.path()
-                )
-            }
-
-            if let Some(query) = self.query.as_ref() {
-                assert_eq!(
-                    query,
-                    request.url.query().unwrap_or_default(),
-                    "[Request {}]: Expected query {}, got {}",
-                    index,
-                    query,
-                    request.url.query().unwrap_or_default()
-                )
-            }
-
-            if let Some(body) = self.body.as_ref() {
-                assert_eq!(
-                    body,
-                    &request.body_json::<serde_json::Value>().unwrap(),
-                    "[Request {}]: incorrect body",
-                    index,
-                )
-            }
-
-            for (name, expected) in self.headers.iter() {
-                match request.headers.get(name) {
-                    Some(actual) => {
-                        let expected: HashSet<String> =
-                            expected.iter().map(|v| v.as_str().to_owned()).collect();
-                        let actual: HashSet<String> =
-                            actual.iter().map(|v| v.as_str().to_owned()).collect();
-                        assert_eq!(
-                            expected,
-                            actual,
-                            "[Request {}]: expected header {} to be [{}], was [{}]",
-                            index,
-                            name,
-                            expected.iter().join(", "),
-                            actual.iter().join(", ")
-                        );
-                    }
-                    None => {
-                        panic!("[Request {}]: expected header {}, was missing", index, name);
-                    }
-                }
-            }
-        }
-    }
-
-    pub(crate) fn matches(received: &[wiremock::Request], matchers: Vec<Matcher>) {
-        assert_eq!(
-            received.len(),
-            matchers.len(),
-            "Expected {} requests, recorded {}",
-            matchers.len(),
-            received.len()
-        );
-        for (i, (request, matcher)) in received.iter().zip(matchers.iter()).enumerate() {
-            matcher.matches(request, i);
-        }
-    }
 }
